feat: add queen position array to support fix of unintentional highlight clear
When two or more queens are already on the board,
toggling one of them to delete will clear the cell
highlights. However, some of those cells are still
within the scope of other queens still existing on
the board, so an array of positions is needed so
the highlights can be cleared and be rebuilt again
among remaining queens.