fix: use querySelector for correct queen detection and adding
The older way of checking queen presence on cells
using hasChildNodes fails to consider the mark text
added, which prevents another queen to be placed
along the scope of prior ones. The mark on the
specific cell clicked should also be removed first
before adding the icon to avoid simultaneous
existence in cells.